Added year tracking to tournament object and reorganized project structure
