+++ objects/Result.py
