+++ objects/Tournament.py
@@ -1,11 +1,12 @@
-from Team import Team
-from TournamentResult import TournamentResult
+from objects.Team import Team
+from objects.TournamentResult import TournamentResult
 
 class Tournament(object):
 
-    def __init__(self, location : str, tournamentType : int):
+    def __init__(self, location : str, tournamentType : int, year : int):
         self.location = location
         self.tournamentType = tournamentType
+        self.year = year
         self.results = []
 
     def addResult(self, rank : int, points : int, team : Team):
@@ -19,6 +20,9 @@ class Tournament(object):
     
     def getTournamentType(self):
         return self.tournamentType
+    
+    def getYear(self):
+        return self.year
         
     def printTournament(self):
         print(f'Location: {self.location}\n')
@@ -28,5 +32,7 @@ class Tournament(object):
 
 def equals(tournament1 : Tournament, tournament2 : Tournament):
     if tournament1.getLocation() != tournament2.getLocation():
+        return False
+    if tournament1.getYear() != tournament2.getYear():
         return False
     return True
